fixes compilation errors on WIN32 (MSVC12)

+++ main/include/eudaq/Utils.hh
@@ -14,6 +14,7 @@
 #include <fstream>
 #include <sys/types.h>
 #include "eudaq/Platform.hh"
+#include <cstdint>
 
 namespace eudaq {
 
@@ -122,11 +123,11 @@ namespace eudaq {
     uint64_t DLLEXPORT from_string(const std::string & x, const uint64_t & def);
   template<>
     inline int32_t DLLEXPORT from_string(const std::string & x, const int32_t & def) {
-      return from_string(x, (int64_t)def);
+      return static_cast<int32_t>(from_string(x, (int64_t)def));
     }
   template<>
     inline uint32_t from_string(const std::string & x, const uint32_t & def) {
-      return from_string(x, (uint64_t)def);
+      return static_cast<uint32_t>(from_string(x, (uint64_t)def));
     }
 
   template <typename T>
